do_findmain

Cluster annotation function: inference of 'M' - molecular weight of the compound giving rise to each spectrum - using the InterpretMSSpectrum::findMain function

do_findmain(
  ramclustObj = NULL,
  cmpd = NULL,
  mode = "positive",
  mzabs.error = 0.005,
  ppm.error = 10,
  ads = NULL,
  nls = NULL,
  scoring = "auto",
  plot.findmain = TRUE,
  writeMat = TRUE,
  writeMS = TRUE,
  use.z = TRUE
)

Arguments

ramclustObj
ramclustR object to annotate.

cmpd
integer: vector defining compound numbers to annotated. if NULL (default), all compounds

mode
character: "positive" or "negative"

mzabs.error
numeric: absolute mass deviation allowd, default = 0.01

ppm.error
numeric: ppm mass error _added_ to mzabs.error, default = 10

ads
character: vector of allowed adducts, i.e. c("[M+H]+"). if NULL, default positive mode values of H+, Na+, K+, and NH4+, as monomer, dimer, and trimer, are assigned. Negative mode include "[M-H]-", "[M+Na-2H]-", "[M+K-2H]-", "[M+CH2O2-H]-" as monomer, dimer, and trimer.

nls
character: vector of allowed neutral losses, i.e. c("[M+H-H2O]+"). if NULL, an extensive list derived from CAMERA's will be used.

scoring
character: one of 'imss' , 'ramclustr', or 'auto'. default = 'auto'. see details.

plot.findmain
logical: should pdf polts be generated for evaluation? detfault = TRUE. PDF saved to working.directory/spectra

writeMat
logical: should indidual .mat files (for MSFinder) be generated in a 'mat' subdirectory in the 'spectra' folder? default = TRUE.

writeMS
logical: should indidual .ms files (for Sirius) be generated in a 'ms' subdirectory in the 'spectra' folder? default = TRUE. Note that no import functions are yet written for Sirius output.

use.z
logical: if you have previously run the 'assign.z' function from ramclustR, there will be a slot reflecting the feature mass after accounting for charge (fm) - if TRUE this is used instead of feature m/z (fmz) in interpreting MS data and exporing spectra for annotation.