Filter by retention time

The function converts seconds into minutes to cut and keep a range of the retention time in minutes.

lcms_filterRTmin(lcms_dataset, rt = c(4, 14))

Arguments

lcms_dataset
An lcms_dataset_family object

rt
Range of the retention time to keep in minutes

Value

A filtered dataset with the range of selected rt

See also

Examples

rtime_range = c(4,14)
lcms_dataset <- lcms_dataset_load(system.file("extdata","lcms_dataset_metadata.rds",package = "NIHSlcms"))
lcms_dataset_rt <-lcms_filterRTmin(lcms_dataset, rt = rtime_range)

print(lcms_dataset_rt)

#> MSn experiment data ("OnDiskMSnExp")
#> Object size in memory: 1.1 Mb
#> - - - Spectra data - - -
#>  MS level(s): 1 
#>  Number of spectra: 4131 
#>  MSn retention times: 4:0 - 13:60 minutes
#> - - - Processing information - - -
#> Data loaded [Tue Oct 15 11:20:19 2019] 
#> Filter: select retention time [240-840] and MS level(s), 1 [Mon Nov 11 16:09:42 2019] 
#>  MSnbase version: 2.8.3 
#> - - - Meta data  - - -
#> phenoData
#>   rowNames: 1 2 3 4
#>   varLabels: sampleNames sampleNumber ... comment (5 total)
#>   varMetadata: labelDescription
#> Loaded from:
#>   [1] 1.mzXML...  [4] 4.mzXML
#>   Use 'fileNames(.)' to see all files.
#> protocolData: none
#> featureData
#>   featureNames: F1.S0358 F1.S0359 ... F4.S1393 (4131 total)
#>   fvarLabels: fileIdx spIdx ... spectrum (30 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'