Filter by sample treatment

In a dataset, there are different types of samples for checking purposes. For instance, quality control (QC) samples may be pools of all samples, and blank samples may be composed by the solvent used. Therefore, these samples should be removed before alignment and processing from the main dataset. Use NULL if there is not any especial sample in the dataset.

lcms_filterSampleType(lcms_dataset, especial_samples)

Arguments

lcms_dataset
An lcms_dataset_family object

especial_samples
A list with the especial samples names. Use NULL if there is not any especial sample in the dataset.

Value

A filtered dataset with the remained samples

Details

The function can distinguish and filter different samples types:

Regular samples

Blank samples

Quality Control samples

See also

Examples

lcms_dataset <- lcms_dataset_load(system.file("extdata","lcms_dataset_metadata.rds",package = "NIHSlcms"))
especial_samples <-list(QC = "resveratrol", blank = "blank")
datasets_by_class_type <- lcms_filterSampleType(lcms_dataset, especial_samples)
lcms_dataset_rs <-datasets_by_class_type$regular_samples
lcms_dataset_qcs <-datasets_by_class_type$QCs

print(lcms_dataset_rs)

#> MSn experiment data ("OnDiskMSnExp")
#> Object size in memory: 2.21 Mb
#> - - - Spectra data - - -
#>  MS level(s): 1 
#>  Number of spectra: 8152 
#>  MSn retention times: 0:31 - 27:2 minutes
#> - - - Processing information - - -
#> Data loaded [Tue Oct 15 11:20:19 2019] 
#> Filter: select file(s) 1, 2, 3. [Fri Oct 18 10:21:46 2019] 
#>  MSnbase version: 2.8.3 
#> - - - Meta data  - - -
#> phenoData
#>   rowNames: 1 2 3
#>   varLabels: sampleNames sampleNumber ... comment (5 total)
#>   varMetadata: labelDescription
#> Loaded from:
#>   [1] 1.mzXML...  [3] 3.mzXML
#>   Use 'fileNames(.)' to see all files.
#> protocolData: none
#> featureData
#>   featureNames: F1.S0001 F1.S0002 ... F3.S2716 (8152 total)
#>   fvarLabels: fileIdx spIdx ... spectrum (30 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'

print(lcms_dataset_qcs)

#> MSn experiment data ("OnDiskMSnExp")
#> Object size in memory: 0.92 Mb
#> - - - Spectra data - - -
#>  MS level(s): 1 
#>  Number of spectra: 2724 
#>  MSn retention times: 0:31 - 27:1 minutes
#> - - - Processing information - - -
#> Data loaded [Tue Oct 15 11:20:19 2019] 
#> Filter: select file(s) 4. [Fri Oct 18 10:21:46 2019] 
#>  MSnbase version: 2.8.3 
#> - - - Meta data  - - -
#> phenoData
#>   rowNames: 4
#>   varLabels: sampleNames sampleNumber ... comment (5 total)
#>   varMetadata: labelDescription
#> Loaded from:
#>   4.mzXML 
#> protocolData: none
#> featureData
#>   featureNames: F4.S0001 F4.S0002 ... F4.S2724 (2724 total)
#>   fvarLabels: fileIdx spIdx ... spectrum (30 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'