Filter by sample type

In a dataset, there are different types of samples for checking purposes. For instance, quality control (QC) samples may be pools of all samples, and blank samples may be composed by the solvent used. Therefore, these samples should be removed before alignment and processing from the main dataset. Use NULL if there is not any especial sample in the dataset.

lcms_filter_sample_type(dataset, especial_samples)

Arguments

dataset
A lcms_dataset

especial_samples
A list with the especial samples names. Use NULL if there is not any especial sample in the dataset.

Value

A list containing three lcms_datasets with different type: regular, quality control and blank samples.

Details

The function can distinguish and filter different samples types:

Regular samples (regular_samples).

Blank samples (blanks).

Quality Control samples (QCs).

See also

Examples

if (FALSE) {
dataset <- lcms_dataset_load(system.file("extdata",
                                        "dataset_pos_rt.rds",
                                        package = "NIHSlcms"))
especial_samples <- list(QC = NULL, blank = NULL)
datasets_by_class_type <- lcms_filter_sample_type(dataset, especial_samples)
dataset_pos_rt_rs <- datasets_by_class_type$regular_samples
dataset_pos_rt_qcs <- datasets_by_class_type$QCs
dataset_pos_rt_bks <- datasets_by_class_type$blanks

print(dataset_pos_rt_rs)
print(dataset_pos_rt_qcs)
print(dataset_pos_rt_bks)
}