Chromatographic peak detection (CentWave)

The findChromPeaks_cwp function performs the chromatographic peak detection on LC/GC-MS data. The standard method for peak detection is 'CentWave'. We must initialize its parameters according to the IPO Package optimization. Peak detection aims to detect important features (peaks) on the chromatographic axis. This will be useful for a posterior peak alignment on the chormatophic axis. Note: signal processing generally consists in three main steps: (1) peak detection (lcms_findChromPeaks_cwp function), (2) peak alignment (lcms_align_Rtime function) and (3) peak correspondence (lcms_group_peaks function).

lcms_findChromPeaks_cwp(lcms_dataset, params)

Arguments

lcms_dataset
An lcms_dataset_family object

params
A converted parameters template from IPO parameters.

Value

A lcms_dataset with detected peaks

Examples

file_path <- system.file("extdata", "rearrange_mait", package = "NIHSlcms")
samples_mzxml <- list.files(file_path, recursive = TRUE, full.names = TRUE)
meta_path <- system.file("extdata", "metadata.xlsx", package = "NIHSlcms")
opt_result_path <-  system.file("extdata", package = "NIHSlcms")
lcms_preproc_params <- lcms_read_ipo_to_xcms(opt_result_path)

lcms_dataset <- suppressWarnings(lcms_read_samples(samples_mzxml, mode = "onDisk"))
metadata <- lcms_meta_read(meta_path)
lcms_dataset_meta <- lcms_meta_add(lcms_dataset, metadata, by = "sampleNames")

peakdet <- lcms_findChromPeaks_cwp(lcms_dataset_meta, params = lcms_preproc_params)

#> 
#>  Finding chromatographic peaks using the optimized set of parameters obtained from IPO package. 

print(peakdet)

#> MSn experiment data ("XCMSnExp")
#> Object size in memory: 3.11 Mb
#> - - - Spectra data - - -
#>  MS level(s): 1 
#>  Number of spectra: 10876 
#>  MSn retention times: 0:31 - 27:2 minutes
#> - - - Processing information - - -
#> Data loaded [Wed Nov 13 18:47:10 2019] 
#>  MSnbase version: 2.10.1 
#> - - - Meta data  - - -
#> phenoData
#>   rowNames: 1 2 3 4
#>   varLabels: sampleNames sampleNumber ... comment (5 total)
#>   varMetadata: labelDescription
#> Loaded from:
#>   [1] 1.mzXML...  [4] 4.mzXML
#>   Use 'fileNames(.)' to see all files.
#> protocolData: none
#> featureData
#>   featureNames: F1.S0001 F1.S0002 ... F4.S2724 (10876 total)
#>   fvarLabels: fileIdx spIdx ... spectrum (33 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> - - - xcms preprocessing - - -
#> Chromatographic peak detection:
#>  method: centWave 
#>  834 peaks identified in 4 samples.
#>  On average 208 chromatographic peaks per sample.