Rearrange datafiles by class

Although Data Preprocessing can be performed using any filepath with the XCMS Package, it is convenient to rearrange data files by sample class (that is, all the samples belonging to the same sample class are included in the same folder). This is done because, after the preprocessing stage, MAIT Package takes care of Data Annotation and Metabolite Identification. MAIT needs a specific directory structure for data managing. Otherwise, it can't work properly.

lcms_rearrange_datafiles_by_class(dataset, dataDir)

Arguments

dataset
A lcms_dataset.

dataDir
A directory in where LC-MS files are going to be saved.

Value

LC-MS datafiles sorted by class treatment.

See also

Examples

if (FALSE) {
file_name <- system.file("extdata", "dataset_pos_rt_rs.rds", package = "NIHSlcms")
dataset <- lcms_dataset_load(file_name)
path <- system.file("extdata","rearrange_mait", package = "NIHSlcms")
lcms_rearrange_datafiles_by_class(dataset = dataset,
                            dataDir = path)
fileList = list.files(path = paste(path, list.files(path = path),
                           sep = "/"), full.names = TRUE)
print(fileList)
}