Significant feature information

Function lcms_sig_peaks_table takes an MAIT-class object containing significant feature information and builds a table with the information related to these features.

lcms_sig_peaks_table(MAIT.object = NULL, printCSVfile = FALSE,
  extendedTable = TRUE, printAnnotation = TRUE)

Arguments

MAIT.object
A MAIT-class object where significant features have already been found.

printCSVfile
Set to TRUE if an output table has to be produced. The table should be found in (working directory)/(project directory)Tables/significativeFeatures.csv.

extendedTable
Set to TRUE the table created by the peak external data is used.

printAnnotation
Set to TRUE The peak annotation is provided in the output table

Value

A table containing: First column (mz): Peak mass Second column (mzmin): Minimum peak mass of the peak group. Third column (mzmax): Maximum peak mass of the peak group. Fourth column (rt): Peak retention time (in minutes). Fifth column (rtmin): Minimum peak retention time of the peak group. Sixth column (rtmax): Maximum peak retention time of the peak group. Seventh column (npeaks): Number of samples where the peak has been detected. The columns from the nineth to the column labeled "isotopes" contain number of class samples where the peak has been detected and the intensities of the peak among samples.

The isotopes column shows if the peak has been identified as a possible isotope.

The adduct column shows which kind of adduct could the peak be.

The column labeled pcgroup contains the spectral ID of the peak.

The P.adjust column contains the corrected peak p-value using post-hoc methods. The p column shows the peak p-value with no multiple test correction.

The Fisher column shows the Fisher test results for the peak. Each of the letters separated by the character "_" corresponds to a class value. Classes having the same letters are indistinguible whereas those having different letters are statistically different clases.

The last columns contain the mean and median values for each feature.

See also

Examples

file_name <-  system.file("extdata", "peak_table_sig_ann.rds", package = "NIHSlcms")
peak_table <- base::readRDS(file_name)
sig_table <- lcms_sig_peaks_table(peak_table,  printCSVfile=FALSE)
str(sig_table)

#> 'data.frame':	642 obs. of  27 variables:
#>  $ mz                             : num  225 225 225 225 225 ...
#>  $ mzmin                          : num  225 225 225 225 225 ...
#>  $ mzmax                          : num  225 225 225 225 225 ...
#>  $ rt                             : num  5.83 5.83 5.83 5.83 5.83 5.83 5.83 5.83 5.84 5.83 ...
#>  $ rtmin                          : num  5.82 5.83 5.81 5.82 5.82 5.82 5.82 5.82 5.82 5.28 ...
#>  $ rtmax                          : num  5.87 5.86 5.86 5.87 5.86 5.86 5.86 5.86 5.86 5.87 ...
#>  $ npeaks                         : num  12 5 7 5 17 9 17 17 17 16 ...
#>  $ plasma                         : num  4 1 4 1 3 4 4 4 4 4 ...
#>  $ plasma_ringer_palm             : num  4 4 2 4 4 4 3 3 4 4 ...
#>  $ X1                             : num  1.75e+07 8.99e+05 1.23e+07 1.80e+06 9.62e+08 ...
#>  $ X2                             : num  4480207 2839559 3119016 648249 99825539 ...
#>  $ X3                             : num  1.05e+07 7.37e+05 7.25e+06 2.32e+05 2.58e+08 ...
#>  $ X4                             : num  4.46e+06 1.20e+05 1.26e+06 3.61e+06 1.12e+08 ...
#>  $ X5                             : num  8.28e+06 5.74e+06 4.27e+05 8.58e+06 1.92e+08 ...
#>  $ X6                             : num  2.77e+07 2.03e+07 2.11e+07 3.00e+07 6.02e+08 ...
#>  $ X7                             : num  5.72e+06 2.95e+06 9.36e+03 5.52e+06 1.41e+08 ...
#>  $ X8                             : num  4.69e+06 1.88e+06 2.34e+06 4.32e+06 1.13e+08 ...
#>  $ isotopes                       : chr  "" "" "" "" ...
#>  $ adduct                         : chr  "" "[M+K]+ 186.159 [M+Na]+ 202.133 [M+NH4]+ 207.089 [M+H]+ 224.115" "[M+K]+ 186.159 [M+Na]+ 202.133 [M+NH4]+ 207.089 [M+H]+ 224.115" "[M+H]+ 224.143" ...
#>  $ pcgroup                        : chr  "1" "1" "1" "1" ...
#>  $ P.adjust                       : num  0.721 0.222 0.996 0.179 0.7 ...
#>  $ p                              : num  0.721 0.222 0.996 0.179 0.7 ...
#>  $ Fisher.Test                    : logi  NA NA NA NA NA NA ...
#>  $ Mean Class plasma              : num  9.23e+06 1.15e+06 5.99e+06 1.57e+06 3.58e+08 ...
#>  $ Mean Class plasma_ringer_palm  : num  1.16e+07 7.71e+06 5.96e+06 1.21e+07 2.62e+08 ...
#>  $ Median Class plasma            : num  7.49e+06 8.18e+05 5.18e+06 1.22e+06 1.85e+08 ...
#>  $ Median Class plasma_ringer_palm: num  7.00e+06 4.34e+06 1.39e+06 7.05e+06 1.66e+08 ...