Build a table of the information related to the significant features contained in a MAIT object

Function sigPeaksTable takes an MAIT-class object containing significant feature information and builds a table with the information related to these features.

sig_peaks_table(MAIT.object = NULL, printCSVfile = FALSE,
  extendedTable = TRUE, printAnnotation = TRUE)

Arguments

MAIT.object
A MAIT-class object where significant features have already been found.

printCSVfile
Set to TRUE if an output table has to be produced. The table should be found in (working directory)/(project directory)Tables/significativeFeatures.csv.

extendedTable
Set to TRUE the table created by the peak external data is used.

printAnnotation
Set to TRUE The peak annotation is provided in the output table

Value

A table containing: First column (mz): Peak mass

Second column(mzmin): Minimum peak mass of the peak group.

Third column(mzmax): Maximum peak mass of the peak group.

Fourth column(rt): Peak retention time (in minutes).

Fifth column(rtmin): Minimum peak retention time of the peak group.

Sixth column(rtmax): Maximum peak retention time of the peak group.

Seventh column(npeaks): Number of samples where the peak has been detected.

The columns from the nineth to the column labeled "isotopes" contain number of class samples where the peak has been detected and the intensities of the peak among samples.

The isotopes column shows if the peak has been identified as a possible isotope.

The adduct column shows which kind of adduct could the peak be.

The column labeled pcgroup contains the spectral ID of the peak.

The P.adjust column contains the corrected peak p-value using post-hoc methods.

The p column shows the peak p-value with no multiple test correction.

The Fisher column shows the Fisher test results for the peak. Each of the letters separated by the character "_" corresponds to a class value. Classes having the same letters are indistinguible whereas those having different letters are statistically different clases.

The last columns contain the mean and median values for each feature

Examples

if (FALSE) {
peak_table_MAIT <- to_MAIT(dataDir = dataDir,
        project = project,
        preproc_params = preproc_params,
        peakTable = peakTable)

peak_table_ANN <- peak_annotation(MAIT.object = peak_table_MAIT)
peak_table_sig_ANN <- spectralSigFeatures(MAIT.object = peak_table_neg_ANN,
pvalue=0.05,
p.adj="none",
scale=FALSE)
sign_table <- sig_peaks_table(peak_table_sig_ANN)
}