Displaying and Storing optimized settings

The function allows visulizing the parameter optimization results by IPO Package allows in the RStudio console. Also you can save this results in plain text files (i.e. a .CVS file).

lcms_write_opt_params(results_pp, results_rtcg, opt_result_path,
  csv = TRUE, console = TRUE)

Arguments

results_pp
object from the lcms_peakpicking_optimizationfunction. If NULL, the default parameters for xcms are loaded.

results_rtcg
object from the lcms_corgroup_optimizationfunction

opt_result_path
A directory to save the parameters file

csv
if TRUE, it writes a file in csv format

console
if TRUE, it displays the params on the console

Value

A file with the params from the IPO optimization

See also

Examples

 opt_result_path <-  system.file("extdata", package = "NIHSlcms")
 file_name_pp <- system.file("extdata", "lcms_resultPeakpicking.rds", package = "NIHSlcms")
 file_name_rcg <- system.file("extdata", "lcms_resultRetcorGroup.rds", package = "NIHSlcms")
 lcms_resultPeakpicking <- base::readRDS(file_name_pp)
 lmcs_resultRetcorGroup <- base::readRDS(file_name_rcg)

 lcms_write_opt_params(lcms_resultPeakpicking, lmcs_resultRetcorGroup, opt_result_path)

#> library(xcms)

#> library(Rmpi)

#> xset <- xcmsSet( 
#>   method = "centWave",
#>   peakwidth       = c(13, 171),
#>   ppm             = 36.95,
#>   noise           = 5000,
#>   snthresh        = 10,
#>   mzdiff          = 0.021,
#>   prefilter       = c(3, 100),
#>   mzCenterFun     = "wMean",
#>   integrate       = 1,
#>   fitgauss        = FALSE,
#>   verbose.columns = FALSE)

#> xset <- retcor( 
#>   xset,
#>   method         = "obiwarp",
#>   plottype       = "none",
#>   distFunc       = "cor_opt",
#>   profStep       = 1,
#>   center         = 4,
#>   response       = 1,
#>   gapInit        = 0.1584,
#>   gapExtend      = 2.7,
#>   factorDiag     = 2,
#>   factorGap      = 1,
#>   localAlignment = 0)

#> xset <- group( 
#>   xset,
#>   method  = "density",
#>   bw      = 0.879999999999999,
#>   mzwid   = 0.0265,
#>   minfrac = 1,
#>   minsamp = 1,
#>   max     = 50)

#> xset <- fillPeaks(xset)

#> $min_peakwidth
#> [1] 13
#> 
#> $max_peakwidth
#> [1] 171
#> 
#> $ppm
#> [1] 36.95
#> 
#> $mzdiff
#> [1] 0.021
#> 
#> $snthresh
#> [1] 10
#> 
#> $noise
#> [1] 5000
#> 
#> $prefilter
#> [1] 3
#> 
#> $value_of_prefilter
#> [1] 100
#> 
#> $mzCenterFun
#> [1] "wMean"
#> 
#> $integrate
#> [1] 1
#> 
#> $fitgauss
#> [1] FALSE
#> 
#> $verbose.columns
#> [1] FALSE
#>